// src/client/src/screens/common/detail.tsx
export const StyledOption = styled.div<{ $width: string, $height: string, $borderRadius: string }>`
    width: ${props => props.$width};
    height: ${props => props.$height};
    background-color: #fff;
    border-radius: ${props => props.$borderRadius};

    display: flex;
    justify-content: center;
    align-items: center;
`
export const StyledOptionButton = styled(StyledOption)`
    &:hover {
        background-color: #fbfbfb;
        cursor: pointer;
    }
    &:active {
        background-color: #f9f9f9;
        transform: scale(0.97);
    }
`